Adds code for thirteenth and first part of fourteenth

+++ utils/TxtReader.cpp
@@ -40,11 +40,11 @@ vector<string> TxtReader::getStringFromFile(string filename) {
   rows.shrink_to_fit();
   return rows;
 }
-int TxtReader::getNumberOfRows(string filename) {
+size_t TxtReader::getNumberOfRows(string filename) {
   ifstream input(filename);
 
   string inputLine;
-  int number = 0;
+  size_t number = 0;
 
   if (input.is_open()) {
     while (getline(input, inputLine)) {
