+++ utils/TxtReader.h
@@ -7,7 +7,7 @@ class TxtReader {
 public:
   std::vector<int> getIntsFromFile(std::string);
   std::vector<std::string> getStringFromFile(std::string);
-  int getNumberOfRows(std::string);
+  size_t getNumberOfRows(std::string);
 };
 
 #endif // TXTREADER_H
